Consolidate templating logic for commands documentation
